+++ gdpr/src/main/java/com/kairos/persistence/repository/master_data/asset_management/data_disposal/DataDisposalMDRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.Set;
 
 @Repository
-public interface DataDisposalMDRepository extends JpaRepository<DataDisposalMD, Integer> {
+public interface DataDisposalMDRepository extends JpaRepository<DataDisposalMD, Long> {
 
     @Query(value = "SELECT d FROM DataDisposalMD d WHERE d.countryId = ?1 and d.deleted = ?2 and lower(d.name) IN ?3")
     List<DataDisposalMD> findByCountryIdAndDeletedAndNameIn(Long countryId, boolean deleted, List<String> userNames);
@@ -22,16 +22,16 @@ public interface DataDisposalMDRepository extends JpaRepository<DataDisposalMD, 
     @Transactional
     @Modifying
     @Query(value = "update DataDisposalMD set name = ?1 where id= ?2")
-    Integer updateDataDisposalName(String name, Integer id);
+    Integer updateDataDisposalName(String name, Long id);
 
 
     @Transactional
     @Modifying
     @Query(value = "update DataDisposalMD set deleted = true where id = ?1 and countryId = ?2")
-    Integer deleteByIdAndCountryId(Integer id, Long countryId);
+    Integer deleteByIdAndCountryId(Long id, Long countryId);
 
     @Query(value = "SELECT d FROM DataDisposalMD d WHERE d.id = ?1 and d.countryId = ?2 and d.deleted = ?3")
-    DataDisposalMD findByIdAndCountryIdAndDeleted(Integer id, Long countryId, boolean deleted);
+    DataDisposalMD findByIdAndCountryIdAndDeleted(Long id, Long countryId, boolean deleted);
 
 
     @Query(value = "SELECT new com.kairos.response.dto.common.DataDisposalResponseDTO(d.id, d.name, d.organizationId, d.suggestedDataStatus, d.suggestedDate )  FROM DataDisposalMD d WHERE d.countryId = ?1 and d.deleted = false order by createdAt desc")
